fix(review): move diagnose-ci step tests to dedicated module with improved assertions
Moves _get_diagnose_ci_content() helper and the two diagnose-ci step-structure
tests out of test_skill_directive_descriptions.py into a new dedicated file
test_diagnose_ci_steps.py. Updates test_diagnose_ci_step_crossref_resolves_correctly
to use re.findall (validates all cross-references) and to assert the full
parenthetical phrase against the heading title rather than individual tokens.
